Created variants.py to fix circular reference issues in schema utils and fixed tests
