+++ app/views/contexts/summary/block.py
@@ -1,6 +1,6 @@
 from flask import url_for
 
-from app.questionnaire.schema_utils import choose_variant
+from app.questionnaire.variants import choose_variant
 from app.views.contexts.summary.question import Question
