+++ app/views/handlers/content.py
@@ -1,6 +1,6 @@
 from functools import cached_property
 
-from app.questionnaire.schema_utils import transform_variants
+from app.questionnaire.variants import transform_variants
 from app.views.handlers import individual_response_url
 from app.views.handlers.block import BlockHandler
