+++ app/views/handlers/question.py
@@ -8,7 +8,7 @@ from app.forms.questionnaire_form import generate_form
 from app.helpers import get_address_lookup_api_auth_token
 from app.questionnaire.location import Location
 from app.questionnaire.questionnaire_store_updater import QuestionnaireStoreUpdater
-from app.questionnaire.schema_utils import transform_variants
+from app.questionnaire.variants import transform_variants
 from app.views.contexts import ListContext
 from app.views.contexts.question import build_question_context
 from app.views.handlers.block import BlockHandler
