+++ tests/app/questionnaire/conftest.py
@@ -11,6 +11,7 @@ from app.questionnaire.location import Location
 from app.questionnaire.placeholder_parser import PlaceholderParser
 from app.questionnaire.placeholder_renderer import PlaceholderRenderer
 from app.questionnaire.placeholder_transforms import PlaceholderTransforms
+from app.questionnaire.rules.operation_helper import OperationHelper
 
 
 @pytest.fixture
@@ -1021,7 +1022,13 @@ def mock_schema():
 
 
 @pytest.fixture
-def placeholder_transform(section_with_labels):
+def placeholder_transform(placeholder):
+    schema, renderer = placeholder
+    return PlaceholderTransforms(language="en", schema=schema, renderer=renderer)
+
+
+@pytest.fixture
+def placeholder(section_with_labels):
     schema = QuestionnaireSchema(section_with_labels)
     answer_id = AnswerStore(
         [
@@ -1037,4 +1044,11 @@ def placeholder_transform(section_with_labels):
         response_metadata={},
         schema=schema,
     )
-    return PlaceholderTransforms(language="en", schema=schema, renderer=renderer)
+    return schema, renderer
+
+
+@pytest.fixture
+def operation_helper(placeholder):
+    schema, renderer = placeholder
+    ops_helper = OperationHelper(language="en", schema=schema, renderer=renderer)
+    return ops_helper
